Practice 02: rpc demo consumer 模拟消费者调用 provider 服务

+++ as-rpc-demo-api/src/main/java/as/rpc/demo/api/OrderService.java
@@ -9,5 +9,5 @@ package as.rpc.demo.api;
  * {@code @date:} 2024/3/19 0:06
  */
 public interface OrderService {
-    Order findById(Long id);
+    Order findById(Integer id);
 }
